Promote boilerplate code to main source set for chapter 10, 11.
